requirements.txt
openai==1.65.5
websockets==10.4
langchain
playwright==1.45.0
langgraph
langchain-openai
aiosqlite 
tavily-python 
loguru
python-dotenv
